+++ .expeditor/scripts/chef_adhoc_build.sh
@@ -4,6 +4,9 @@ set -euo pipefail
 
 hab_target="${1:-x86_64-linux}"
 
+# ensure habitat 2.0.504 is installed
+./.expeditor/scripts/install-hab.sh "$hab_target"
+
 export HAB_ORIGIN='chef'
 export PLAN='chef-infra-client'
 export CHEF_LICENSE="accept-no-persist"