+++ .expeditor/scripts/validate_adhoc_build.sh
@@ -36,7 +36,7 @@ if [ $? -ne 0 ]; then
 fi
 
 echo "--- Installing $PKG_ARTIFACT"
-sudo hab pkg install $PKG_ARTIFACT --auth $HAB_AUTH_TOKEN --binlink
+sudo -E hab pkg install $PKG_ARTIFACT --binlink
 
 pkg_ident=$(hab pkg list "$HAB_ORIGIN"/"$PLAN")
 echo "--- Resolved package identifier: $pkg_ident, attempting to run tests"